+++ tencentcloud/cpdp/v20190820/models.go
@@ -14689,6 +14689,10 @@ type PayeeAccountBalanceResult struct {
 	// 提现中余额
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	InPayBalance *string `json:"InPayBalance,omitempty" name:"InPayBalance"`
+
+	// 累计结算金额
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	SumSettlementAmount *string `json:"SumSettlementAmount,omitempty" name:"SumSettlementAmount"`
 }
 
 type PayeeAccountInfoResult struct {