+++ tencentcloud/dlc/v20210125/errors.go
@@ -77,6 +77,9 @@ const (
 	// 无效的Offset值。
 	INVALIDPARAMETER_INVALIDOFFSET = "InvalidParameter.InvalidOffset"
 
+	// 无效的CAM role arn。
+	INVALIDPARAMETER_INVALIDROLEARN = "InvalidParameter.InvalidRoleArn"
+
 	// SQL解析失败。
 	INVALIDPARAMETER_INVALIDSQL = "InvalidParameter.InvalidSQL"