+++ tencentcloud/oceanus/v20190422/errors.go
@@ -29,6 +29,9 @@ const (
 	// 重复的作业名。
 	FAILEDOPERATION_DUPLICATEDJOBNAME = "FailedOperation.DuplicatedJobName"
 
+	// 查询资源关联标签失败。
+	FAILEDOPERATION_GETRESOURCETAGSBYRESOURCEIDS = "FailedOperation.GetResourceTagsByResourceIds"
+
 	// 语法检查失败。
 	FAILEDOPERATION_GRAMMARCHECKFAILURE = "FailedOperation.GrammarCheckFailure"
 
@@ -119,6 +122,9 @@ const (
 	// 集群Id指定错误，为空或者不存在。
 	INVALIDPARAMETERVALUE_CLUSTERID = "InvalidParameterValue.ClusterId"
 
+	// 群集Id报错。
+	INVALIDPARAMETERVALUE_CLUSTERIDS = "InvalidParameterValue.ClusterIds"
+
 	// CU内存规格不匹配。
 	INVALIDPARAMETERVALUE_CUMEM = "InvalidParameterValue.CuMem"
 
@@ -139,6 +145,9 @@ const (
 
 	// SQL作业不能指定EntrypointClass，JAR作业则必须指定。
 	INVALIDPARAMETERVALUE_JOBTYPECOMBINEWITHENTRYPOINTCLASS = "InvalidParameterValue.JobTypeCombineWithEntrypointClass"
+
+	// OrderType值错误。
+	INVALIDPARAMETERVALUE_ORDERTYPE = "InvalidParameterValue.OrderType"
 
 	// 未找到资源ID。
 	INVALIDPARAMETERVALUE_RESOURCEIDSNOTFOUND = "InvalidParameterValue.ResourceIdsNotFound"
@@ -230,6 +239,9 @@ const (
 	// 找不到JobId的正在运行的作业实例。
 	RESOURCEUNAVAILABLE_NORUNNINGJOBINSTANCESFOUNDFORJOBID = "ResourceUnavailable.NoRunningJobInstancesFoundForJobId"
 
+	// 不允许删除错误。
+	RESOURCEUNAVAILABLE_NOTALLOWEDTOBEDELETED = "ResourceUnavailable.NotAllowedToBeDeleted"
+
 	// 不允许停止或暂停错误。
 	RESOURCEUNAVAILABLE_NOTALLOWEDTOBESTOPORPAUSE = "ResourceUnavailable.NotAllowedToBeStopOrPause"
 
@@ -238,6 +250,9 @@ const (
 
 	// cos上的程序包不存在。
 	RESOURCEUNAVAILABLE_RESOURCELOCNOTEXISTS = "ResourceUnavailable.ResourceLocNotExists"
+
+	// 未授权操作。
+	UNAUTHORIZEDOPERATION = "UnauthorizedOperation"
 
 	// 操作不支持。
 	UNSUPPORTEDOPERATION = "UnsupportedOperation"